Forms for UI

+++ src/main/java/fyp/repo/SiteRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import fyp.entities.Site;
 
 
-// automatic translation between DB (MySQL) and POJO (TimesheetEntry)
-public interface SiteRepository extends JpaRepository<Site, Long> {
+// automatic translation through ORM between MySQL DB and POJO (Site)
+public interface SiteRepository extends JpaRepository<Site, Integer> {
 }
