+++ src/main/java/fyp/repo/TimesheetEntryRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import fyp.entities.TimesheetEntry;
 
 
-// automatic translation between DB (MySQL) and POJO (TimesheetEntry)
-public interface TimesheetEntryRepository extends JpaRepository<TimesheetEntry, Long> {
+// automatic translation between MySQL DB and POJO (TimesheetEntry)
+public interface TimesheetEntryRepository extends JpaRepository<TimesheetEntry, Integer> {
 }
