+++ src/main/java/fyp/repo/UserGroupRepository.java
@@ -2,10 +2,9 @@ package fyp.repo;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import fyp.entities.Site;
 import fyp.entities.UserGroup;
 
 
-// automatic translation between DB (MySQL) and POJO (TimesheetEntry)
-public interface UserGroupRepository extends JpaRepository<UserGroup, Long> {
+// automatic translation through ORM between MySQL DB and POJO (UserGroup)
+public interface UserGroupRepository extends JpaRepository<UserGroup, Integer> {
 }
